Move Vehicle linear movement into Trajectory class (#13)
This is the first step in generalizing pathing/movement to make way for curved trajectories when Vehicles are moving around corners.

The curved trajectory logic appears to be significantly complex, so I wansn't able to get around to it just yet. More information at: https://stackoverflow.com/questions/59783526/move-object-a-set-distance-along-an-elliptical-path?noredirect=1#comment105711071_59783526. Kepler's equation could be a good place to start. Otherwise it's a matter of deconstructing how to find the arc length of an ellipse efficiently. I added more detail to #6 to better explain the options.

Ultimately we could create an EllipticalTrajectory class in physics/pathing.py that will handle the trajectories when a player takes a turn.

Also, I switched my development over to VSCode to try it out. Added some changes like flake8 to support that. Should have been its own PR/commit, but I didn't have the time to split it out.